+++ Tests/Functional/NoTemplatingEntryTest.php
@@ -64,16 +64,9 @@ class NoTemplatingEntryKernel extends Kernel
     {
         $loader->load(function (ContainerBuilder $container) {
             $config = [
-                'annotations' => false,
-                'http_method_override' => false,
-                'php_errors' => ['log' => true],
                 'secret' => '$ecret',
                 'form' => ['enabled' => false],
             ];
-
-            if (trait_exists(HttpClientAssertionsTrait::class)) {
-                $config['handle_all_throwables'] = true;
-            }
 
             $container
                 ->loadFromExtension('framework', $config)