+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Make `TemplateCacheWarmer` class `final`
+ * Remove the `base_template_class` config option
 
 7.3
 ---